move typed DataContexts into public API
